fix(proxy): address Greptile review feedback
1. Defer SSE request.started event until after modelID is resolved.
   Previously insertRequestLogAsync published the event with an empty
   model_id when called before body parsing. Now publishRequestStartedEvent
   is called after modelID is set, and also before early-return failure
   paths so subscribers always see the event.

2. Extract isCancelled into shared web/src/utils/logHelpers.ts to
   eliminate drift risk between Logs.tsx and VirtualLogTable.tsx.
